fix(date-range): when dates are null minmax values are not enforced
When  I removed the start/end dates and used the year/month calendar views to set the start date to
a day in the future (which I was able to) and set the end date to "today's date". The "Apply" button
is enabled. In this scenario, I was expecting the "Apply" button to be

fix #763