fix(date-range): allow blank dates when to/from dates aren't required
A date range picker where 'to date required' and 'from date required' are set to `false` should allow a completely empty range with two null values for the date to be created. This change amends the range validation function to support the aforementioned scenario and adds a test to verify it.

Closes #1833