fix(examples): fix rxjs-triggered build failures
We've been seeing obscure rxjs-related errors when trying to build and run cashmere-examples. It
appears that these are caused by the presence of two different versions of rxjs within our project.
This prospective fix seeks to resolve this issue by informing cashmere-examples that it should
always use the version 6 dependency. Note: this change should be removed as part of the Angular 13
upgrade, as that will bring us up to rxjs 7 across the board.