fix(drawer): prevent drawer from getting 'stuck'
Immediately return a resolved Promise from the Drawer component's `toggle()` function if the desired state of the Drawer matches the current state of the Drawer. This will prevent the Drawer from getting 'stuck' open when an open Drawer is asked to open itself.

This allows callers to request a particular state without worrying about the Drawer's current state.

fix #2005