chore(package.json): allow angular cdk peer greater than current version
Currently, we have a section on the Cashmere website suggesting that consumer Angular versions do
not need to track the version used by Cashmere, but Angular CDK requires a major version match,
while the rest do not. Without this, it's not possible, in practice, to upgrade to a newer version
of Angular until Cashmere is also updated.